Allow to ask caller to push needed images during fixup
During the `fixup` phase, we ensure that all images are available
before to patch it with digest.
Until now, all needed images needed to be pushed before to run the fixup.

The fixup options now contains a `pushImageCallback` field. If defined,
this callback will allow the caller to push the needed and not found image
and will answer with the digest. This way it's not needed to push images
before, they will be pushed on demand.

If the callback is not set, the behavior remains the same so an error is
raised and the fixup fails.

Signed-off-by: Yves Brissaud <yves.brissaud@docker.com>